MAINT: restructure source files
