+++ src/core/parser.ts
